pyproject.toml
[project]
name = "resume-ats-score-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.5",
    "camelot-py>=1.0.9",
    "fastapi[standard]>=0.116.1",
    "firecrawl-py>=4.1.1",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jsonschema>=4.25.1",
    "langchain-google-genai>=2.1.10",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pymongo>=4.14.1",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-box>=7.3.2",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "selenium>=4.35.0",
    "toml>=0.10.2",
]

[build-system]
requires = [ "hatchling >= 1.27.0",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = [ "src/ats",]
